+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.6.0-dev
 
 - Additional formatting directives
+  - `%m` - Request method
   - `%{VARNAME}o` - Header line sent by the server
 
 ## v0.5.0 (2015-02-24)